pyproject.toml
[project]
name = "tf-module-updater"
version = "0.1.0"
description = "Tool for updating Terraform modules efficiently."
authors = [{email="jacob.verhoeks@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "packaging>=25.0",
    "python-hcl2>=7.2.1",
    "requests>=2.32.4",
]

[tool]
uv.package = true

[project.scripts]
tf-module-updater = "tf_module_updater:main"

